Avoid duplicate entries
Introduced the containsObject function that checks if an object is in an array. Now this is non trivial, because even if a and b are identical objects, a==b or a===b returns false. only a==a is true. So the function now compares all the properties and then returns true (object is present in the array) if the counter matches the length of the object.